Dispatch: document MAP blocks

+++ iMUSE/tracks.c
@@ -119,7 +119,7 @@ void tracks_callback() {
 	waveapi_increaseSlice();
 	dispatch_predictFirstStream();
 	if (tracks_waveCall) {
-		waveapi_write((int)&iMUSE_audioBuffer, (int)&iMUSE_feedSize, iMUSE_sampleRate);
+		waveapi_write(&iMUSE_audioBuffer, &iMUSE_feedSize, iMUSE_sampleRate);
 	} else {
 		// 40 Hz frequency for filling the audio buffer, for some reason
 		// Anyway it appears we never reach this block since tracks_waveCall is assigned to a (dummy) function
@@ -162,7 +162,7 @@ void tracks_callback() {
 		if (tracks_waveCall) {
 			mixer_loop(iMUSE_audioBuffer, iMUSE_feedSize);
 			if (tracks_waveCall) {
-				waveapi_write((int)&iMUSE_audioBuffer, (int)&iMUSE_feedSize, 0);
+				waveapi_write(&iMUSE_audioBuffer, &iMUSE_feedSize, 0);
 			}
 		}
 	}
